refactor: code cleanup from main merge
