+++ package.json
@@ -46,6 +46,7 @@
     "@sveltejs/kit": "1.0.0-next.297",
     "@types/chroma-js": "^2.1.3",
     "@types/colormap": "^2.3.1",
+    "@types/lodash": "^4.14.180",
     "@types/papaparse": "^5.3.2",
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.15.0",
